SAK-49897 conversations. Honour cleanup flag in site import
https://sakaiproject.atlassian.net/browse/SAK-49897

+++ conversations/impl/src/main/java/org/sakaiproject/conversations/impl/ConversationsServiceImpl.java
@@ -2333,13 +2333,16 @@ public class ConversationsServiceImpl implements ConversationsService, EntityPro
 
     public Map<String, String> transferCopyEntities(String fromContext, String toContext, List<String> ids, List<String> transferOptions, boolean cleanup) {
 
-        topicRepository.findBySiteId(toContext).forEach(t -> {
+        if (cleanup) {
+            topicRepository.findBySiteId(toContext).forEach(t -> {
 
-            try {
-                deleteTopic(t.getId());
-            } catch (ConversationsPermissionsException cpe) {
-            }
-        });
+                try {
+                    deleteTopic(t.getId());
+                } catch (ConversationsPermissionsException cpe) {
+                    log.warn("No permission to delete topic {}", t.getId());
+                }
+            });
+        }
 
         return transferCopyEntities(fromContext, toContext, ids, transferOptions);
     }
